+++ cloneholio/github.py
@@ -7,7 +7,7 @@ import github
 LOGGER = logging.getLogger('cloneholio')
 
 
-def get_repos(path, token, insecure=False, base_url=None):
+def get_repos(path, token, insecure=False, base_url=None, archived=True):
     kwargs = {'verify': not insecure}
     if base_url:
         kwargs['base_url'] = base_url
@@ -33,6 +33,8 @@ def get_repos(path, token, insecure=False, base_url=None):
         repos = api.get_user(path).get_repos()
 
     for repo in repos:
+        if repo.archived and not archived:
+            continue
         yield repo.full_name, repo.ssh_url